requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
crispy-bootstrap5==0.6
defusedxml==0.7.1
Django==4.0.5
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
djangorestframework==3.13.1
httpie==3.2.1
idna==3.3
multidict==6.0.2
Pygments==2.12.0
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.4.4
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
